canid needs 32 bits

+++ core/USBInterface/USBInterface.cpp
@@ -84,7 +84,7 @@ void USBPort::handle()
    uint bytesread;
    size_t findhash,findinfo;
    uint16_t port;
-   uint16_t canid;
+   uint32_t canid;
    struct can_frame frame;
    char cbuf[2048];
 
@@ -147,7 +147,7 @@ void USBPort::handle()
          findhash = sbuf.find_first_of('#');
          if (findhash == HASH_OFFSET && m_unprocessedSize >= STD_MSG_SIZE)
          {
-            // msg format 0P000xxx P=port xxx=msg id
+            // OLD msg format 0P000xxx P=port xxx=msg id
             // msg format 0Pxxxxxxxx#yyyyyyyyyyyyyyyy P=port xxxxxxxx=can id (8 bytes) hex y..=can msg hex (8 bytes)
             // get the can port number from the long id
             port = HextoDec( &m_inBufferUnprocessed[PORT_OFFSET], 1);
